# packages/schemas/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "partial",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "Liveness": {
        "properties": {
          "service": {
            "const": "mask-api",
            "default": "mask-api",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "Liveness",
        "type": "object"
      },
      "Readiness": {
        "properties": {
          "dependencies": {
            "additionalProperties": {
              "enum": [
                "up",
                "down"
              ],
              "type": "string"
            },
            "propertyNames": {
              "const": "postgres"
            },
            "title": "Dependencies",
            "type": "object"
          },
          "status": {
            "enum": [
              "ready",
              "not_ready"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "dependencies"
        ],
        "title": "Readiness",
        "type": "object"
      },
      "SmokeRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "format": "uuid",
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key"
        ],
        "title": "SmokeRequest",
        "type": "object"
      },
      "SmokeResponse": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "execution_count": {
            "title": "Execution Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "id",
          "status",
          "correlation_id",
          "execution_count",
          "attempt_count",
          "created_at"
        ],
        "title": "SmokeResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MASK Infrastructure API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/dev/jobs/smoke": {
      "post": {
        "operationId": "smoke_dev_jobs_smoke_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Dev-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmokeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Smoke"
      }
    },
    "/dev/jobs/{job_id}": {
      "get": {
        "operationId": "job_status_dev_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dev-token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Dev-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmokeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Status"
      }
    },
    "/health/live": {
      "get": {
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Liveness"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Live"
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ready"
      }
    }
  }
}